+++ Sources/MapboxNavigation/Feedback/FeedbackItemType.swift
@@ -1,0 +1,27 @@
+import MapboxCoreNavigation
+
+public enum FeedbackItemType: Equatable {
+    case activeNavigation(FeedbackType)
+    case passiveNavigation(PassiveNavigationFeedbackType)
+    
+    public static func == (lhs: FeedbackItemType, rhs: FeedbackItemType) -> Bool {
+        switch (lhs, rhs) {
+        case (.activeNavigation, .passiveNavigation),
+             (.passiveNavigation, .activeNavigation):
+            return false
+        case (.activeNavigation(let leftType), .activeNavigation(let rightType)):
+            return leftType.title == rightType.title
+        case (.passiveNavigation(let leftType), .passiveNavigation(let rightType)):
+            return leftType.title == rightType.title
+        }
+    }
+    
+    var title: String {
+        switch self {
+        case .activeNavigation(let type):
+            return type.title
+        case .passiveNavigation(let type):
+            return type.title
+        }
+    }
+}
